test/requirements.txt
openai>=1.65.2
pymupdf>=1.23.0
python-dotenv>=1.0.0
